src/brain_of_droid/brain_of_droid/brain/requirements.txt
python-dotenv
langgraph
langchain-core
langchain-openai
langchain-community
langchain-chroma
requests
pyaudio
pyttsx3
openai
python-dotenv